+++ docs/release-notes.md
@@ -14,6 +14,7 @@
 
 * For mod authors:
   * Added `World.FurnitureListChanged` event (thanks to DiscipleOfEris!).
+  * Added asset propagation for building/house paint masks.
   * Added validation for the manifest `Dependencies` field.
   * When using Visual Studio Code to edit JSON files [with a SMAPI JSON schema](technical/web.md#using-a-schema-file-directly), it will no longer warn about comments or trailing commas.
   * Fixed validation for mods with version `0.0.0`.